fix ui endpoint in kpi package
