--- extensions/wrapper/client-ts/README.md
@@ -1,27 +1,0 @@
-<!-- PROJECT LOGO -->
-<br />
-<div align="center">
-  <a href="https://github.com/sovity/edc-extensions">
-    <img src="https://raw.githubusercontent.com/sovity/edc-ui/main/src/assets/images/sovity_logo.svg" alt="Logo" width="300">
-  </a>
-
-<h3 align="center">EDC-Connector Extension:<br />API Wrapper &amp; API Client:<br />TypeScript API Client</h3>
-
-  <p align="center">
-    <a href="https://github.com/sovity/edc-extensions/issues/new?template=bug_report.md">Report Bug</a>
-    ·
-    <a href="https://github.com/sovity/edc-extensions/issues/new?template=feature_request.md">Request Feature</a>
-  </p>
-</div>
-
-## About this component
-
-TypeScript Client Library to be imported and used in arbitrary applications like frontends or nodejs projects.
-
-## License
-
-Apache License 2.0 - see [LICENSE](../../LICENSE)
-
-## Contact
-
-sovity GmbH - contact@sovity.de
