+++ extensions/wrapper/wrapper/src/main/java/de/sovity/edc/ext/wrapper/api/ui/model/AssetDto.java
@@ -12,7 +12,7 @@
  *
  */
 
-package de.sovity.edc.ext.wrapper.api.usecase.model;
+package de.sovity.edc.ext.wrapper.api.ui.model;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
